reorg of guards.ts into primitives
